check for valid ID before trying to delete nutrition data

+++ src/main/java/com/health/MinimalismFitnessApp/controllers/NutritionController.java
@@ -4,6 +4,7 @@ import com.health.MinimalismFitnessApp.entities.NutritionData;
 import com.health.MinimalismFitnessApp.services.NutritionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -42,8 +43,15 @@ public class NutritionController {
     }
 
     @DeleteMapping("/{nutritionID}")
-    public void deleteNutritionData(@PathVariable long nutritionID){
-        nutritionService.deleteNutritionData(nutritionID);
+    public ResponseEntity<Long> deleteNutritionData(@PathVariable long nutritionID){
+        boolean validID = nutritionService.checkNutritionDataExistsByID(nutritionID);
+
+        if (validID){
+            nutritionService.deleteNutritionData(nutritionID);
+            return new ResponseEntity<>(nutritionID, HttpStatus.OK);
+        } else{
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
 }
